CURB-4179 Introduced AppPermissionRequest class to implement app commands with corresponding response revents
- new class is based on the AppPermissionsRequest classed
- refactored AppPermissionsRequest to extend AppPermissionRequest
